+++ pkgs/cupertino_http/ios_test/integration_test/combined_test.dart
@@ -1,0 +1,12 @@
+// Copyright (c) 2026, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:integration_test/integration_test.dart';
+
+import '../../test/main.dart' as test_main;
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  test_main.main();
+}